pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "instamatte"
version = "0.3.0"
description = "Multi-format image processor for social media"
authors = [{name = "Arthur Ryzhak", email = "ryzhakar@gmail.com"}]
requires-python = ">=3.9"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "Pillow>=10.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
instamatte = "instamatte.cli:cli_application"

[tool.ruff]
line-length = 80
target-version = "py39"
